Calculate MFC

Calculate_MFC calculates the (log-transformed) maximum fold change over all strains.

Calculate_MFC(dat_list, subjectCol = "SubjectID", discretize = c(0.2, 0.3),
  responseLabels = paste0(c("low", "moderate", "high"), "Responder"))

Arguments

dat_list
a named list like the one returned by FormatTiters.

subjectCol
the name of the column specifying a subject ID. Default is "SubjectID".

discretize
a vector of quantiles in (0, 0.5] specifying where to make the cutoff for low, moderate and high responses. Default is 20% and 30%.

responseLabels
names for low, moderate and high responses

Value

A list with the following elements:

MFC

a named vector containing the continuous MFC endpoints

MFC_d<X>

a named vector containing the discrete MFC endpoint with a cutoff at <X>

...

Other named vectors containing discrete MFC endpoints

A named list containing the MFC for each subject and any discretized metrics

Examples

## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

#> - Log transforming Pre and Post columns

#> - Setting any negative log fold changes to 0

Calculate_MFC(titer_list)

#> $MFC
#> 110196 110264 110281 110266 110278 110285 110193 110198 110210 110280 110283 
#>      1      1      0      4      1      1      2      2      2      3      3 
#> 110259 110216 110267 110268 110247 110207 110208 110257 110197 110199 110214 
#>      5      6      6      7      8      1      2      4      5      1      0 
#> 110202 110205 110271 110282 110284 110195 110201 110255 110277 110265 110272 
#>      1      2      1      1      1      2      3      3      4      6      6 
#> 110192 110209 110215 110275 110194 110203 110213 110248 110260 110263 110204 
#>      0      0      1      2      1      2      2      1      1      1      2 
#> 110211 110241 110256 110276 110243 110252 110274 110286 110191 110212 110245 
#>      3      3      4      3      1      3      1      1      1      1      3 
#> 110250 110262 110254 110269 110251 110270 110279 110200 110261 110273 110244 
#>      1      3      2      2      3      0      0      2      2      1      3 
#> 110258 110242 110253 
#>      1      1      1 
#> 
#> $MFC_d20
#>            110196            110264            110281            110266 
#>      lowResponder      lowResponder      lowResponder     highResponder 
#>            110278            110285            110193            110198 
#>      lowResponder      lowResponder moderateResponder moderateResponder 
#>            110210            110280            110283            110259 
#> moderateResponder     highResponder     highResponder     highResponder 
#>            110216            110267            110268            110247 
#>     highResponder     highResponder     highResponder     highResponder 
#>            110207            110208            110257            110197 
#>      lowResponder moderateResponder     highResponder     highResponder 
#>            110199            110214            110202            110205 
#>      lowResponder      lowResponder      lowResponder moderateResponder 
#>            110271            110282            110284            110195 
#>      lowResponder      lowResponder      lowResponder moderateResponder 
#>            110201            110255            110277            110265 
#>     highResponder     highResponder     highResponder     highResponder 
#>            110272            110192            110209            110215 
#>     highResponder      lowResponder      lowResponder      lowResponder 
#>            110275            110194            110203            110213 
#> moderateResponder      lowResponder moderateResponder moderateResponder 
#>            110248            110260            110263            110204 
#>      lowResponder      lowResponder      lowResponder moderateResponder 
#>            110211            110241            110256            110276 
#>     highResponder     highResponder     highResponder     highResponder 
#>            110243            110252            110274            110286 
#>      lowResponder     highResponder      lowResponder      lowResponder 
#>            110191            110212            110245            110250 
#>      lowResponder      lowResponder     highResponder      lowResponder 
#>            110262            110254            110269            110251 
#>     highResponder moderateResponder moderateResponder     highResponder 
#>            110270            110279            110200            110261 
#>      lowResponder      lowResponder moderateResponder moderateResponder 
#>            110273            110244            110258            110242 
#>      lowResponder     highResponder      lowResponder      lowResponder 
#>            110253 
#>      lowResponder 
#> Levels: lowResponder moderateResponder highResponder
#> 
#> $MFC_d30
#>            110196            110264            110281            110266 
#>      lowResponder      lowResponder      lowResponder     highResponder 
#>            110278            110285            110193            110198 
#>      lowResponder      lowResponder moderateResponder moderateResponder 
#>            110210            110280            110283            110259 
#> moderateResponder     highResponder     highResponder     highResponder 
#>            110216            110267            110268            110247 
#>     highResponder     highResponder     highResponder     highResponder 
#>            110207            110208            110257            110197 
#>      lowResponder moderateResponder     highResponder     highResponder 
#>            110199            110214            110202            110205 
#>      lowResponder      lowResponder      lowResponder moderateResponder 
#>            110271            110282            110284            110195 
#>      lowResponder      lowResponder      lowResponder moderateResponder 
#>            110201            110255            110277            110265 
#>     highResponder     highResponder     highResponder     highResponder 
#>            110272            110192            110209            110215 
#>     highResponder      lowResponder      lowResponder      lowResponder 
#>            110275            110194            110203            110213 
#> moderateResponder      lowResponder moderateResponder moderateResponder 
#>            110248            110260            110263            110204 
#>      lowResponder      lowResponder      lowResponder moderateResponder 
#>            110211            110241            110256            110276 
#>     highResponder     highResponder     highResponder     highResponder 
#>            110243            110252            110274            110286 
#>      lowResponder     highResponder      lowResponder      lowResponder 
#>            110191            110212            110245            110250 
#>      lowResponder      lowResponder     highResponder      lowResponder 
#>            110262            110254            110269            110251 
#>     highResponder moderateResponder moderateResponder     highResponder 
#>            110270            110279            110200            110261 
#>      lowResponder      lowResponder moderateResponder moderateResponder 
#>            110273            110244            110258            110242 
#>      lowResponder     highResponder      lowResponder      lowResponder 
#>            110253 
#>      lowResponder 
#> Levels: lowResponder moderateResponder highResponder
#>