Calculate Nakaya2015

Calculate_Nakaya2015 calculates the endpoint used in Nakaya et al. 2015

Calculate_Nakaya2015(dat_list, subjectCol = "SubjectID",
  responseLabels = paste0(c("low", "high"), "Responder"),
  na_action = "na.fail", ...)

Arguments

dat_list
a named list like the one returned by FormatTiters.

subjectCol
the name of the column specifying a subject ID. Default is "SubjectID".

responseLabels
names for low and high responses

na_action
how should missing NA values be treated. Default is "na.fail"

...
Additional arguments passed to lm

Value

A list with the following elements:

data

a data frame containing the MFC and indicator variables that determine whether subject is a low or high responder (see details)

Nakaya2015

a named vector containing the discretized endpoint

Details

First calculate the maximum fold change (MFC) derived titer metric described in Nakaya et al. 2015. Then check whether both of these conditions are satisfied: i) MFC is at least a 4-fold increase ii) The "Post" antibody titer is 1:40 or more for at least 1 strain Subjects are classified as high responders if they satisfy both conditions and low responders otherwise.

Missing (NA) values are handled by being returned as missing in the endpoints in the output

References

Nakaya HI, et al. (2015) Systems Analysis of Immunity to Influenza Vaccination across Multiple Years and in Diverse Populations Reveals Shared Molecular Signatures. Immunity 43(6):1186-1198.

See also

CalculateMFC

Examples

## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

#> - Log transforming Pre and Post columns

#> - Setting any negative log fold changes to 0

## Calculate the endpoint
endpoints <- Calculate_Nakaya2015(titer_list)
summary(endpoints)

#>            Length Class      Mode     
#> data        5     data.frame list     
#> Nakaya2015 69     -none-     character