Calculate Thakar2015

Calculate_Thakar2015 calculates a response definition based on a Four Fold change to At Least Two Strains (Thakar2015).

Calculate_Thakar2015(dat_list, subjectCol = "SubjectID",
  responseLabels = c("NR", "X", "R"))

Arguments

dat_list
a named list like the one returned by FormatTiters.

subjectCol
the name of the column specifying a subject ID. Default is "SubjectID".

responseLabels
names for low, middle, and high responses

Value

A named list with 1 element named "Thakar2015" containing the response ("NR", "X", or "R").

Details

Subjects are responders (default "R") if they acheive a 4-fold or greater fold change in titer to at least 2 strains, nonresponders (default "NR") if they do not acheive a 4-fold or greater fold change in titer to any strain, and intermediate (default "X") otherwise. Missing (NA) values are handled by being returned as missing in the endpoints in the output

References

Thakar J, et al. (2015) Aging-dependent alterations in gene expression and a mitochondrial signature of responsiveness to human influenza vaccination. Aging (Albany NY) 7(1):38-52.

Examples

## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

#> - Log transforming Pre and Post columns

#> - Setting any negative log fold changes to 0

Calculate_Thakar2015(titer_list)

#> $Thakar2015
#> 110191 110192 110193 110194 110195 110196 110197 110198 110199 110200 110201 
#>   "NR"   "NR"    "X"   "NR"    "R"   "NR"    "R"    "X"   "NR"    "X"    "R" 
#> 110202 110203 110204 110205 110207 110208 110209 110210 110211 110212 110213 
#>   "NR"    "X"    "R"    "R"   "NR"    "X"   "NR"    "X"    "X"   "NR"    "X" 
#> 110214 110215 110216 110241 110242 110243 110244 110245 110247 110248 110250 
#>   "NR"   "NR"    "R"    "X"   "NR"   "NR"    "X"    "X"    "R"   "NR"   "NR" 
#> 110251 110252 110253 110254 110255 110256 110257 110258 110259 110260 110261 
#>    "X"    "X"   "NR"    "X"    "X"    "R"    "R"   "NR"    "R"   "NR"    "X" 
#> 110262 110263 110264 110265 110266 110267 110268 110269 110270 110271 110272 
#>    "X"   "NR"   "NR"    "R"    "R"    "R"    "X"    "X"   "NR"   "NR"    "R" 
#> 110273 110274 110275 110276 110277 110278 110279 110280 110281 110282 110283 
#>   "NR"   "NR"    "X"    "X"    "R"   "NR"   "NR"    "X"   "NR"   "NR"    "R" 
#> 110284 110285 110286 
#>   "NR"   "NR"   "NR" 
#>