Calculate pre-GMT

Calculate_preGMT calculates the log-transformed pre-vaccination geometric mean titer (pre-GMT)

Calculate_preGMT(dat_list, subjectCol = "SubjectID")

Arguments

dat_list
a named list like the one returned by FormatTiters.

subjectCol
the name of the column specifying a subject ID. Default is "SubjectID".

Value

A named vector containing the pre-GMT for each subject

Details

Non-logged HAI titers for each strain are used to calculate the geometric mean and the geometric mean for each subject is subsequently log2-transformed.

Examples

## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

#> - Log transforming Pre and Post columns

#> - Setting any negative log fold changes to 0

Calculate_preGMT(titer_list)

#>   110196   110264   110281   110266   110278   110285   110193   110198 
#> 2.000000 2.333333 3.000000 2.333333 2.000000 3.666667 2.666667 2.000000 
#>   110210   110280   110283   110259   110216   110267   110268   110247 
#> 2.000000 2.333333 2.666667 3.666667 3.000000 3.000000 3.666667 2.000000 
#>   110207   110208   110257   110197   110199   110214   110202   110205 
#> 2.666667 2.666667 3.666667 2.333333 3.666667 4.666667 4.666667 3.666667 
#>   110271   110282   110284   110195   110201   110255   110277   110265 
#> 4.000000 6.000000 4.333333 3.666667 3.333333 4.666667 4.000000 2.666667 
#>   110272   110192   110209   110215   110275   110194   110203   110213 
#> 2.666667 5.000000 4.666667 4.333333 4.000000 4.333333 4.000000 3.666667 
#>   110248   110260   110263   110204   110211   110241   110256   110276 
#> 4.666667 4.000000 4.333333 4.333333 5.000000 5.333333 3.666667 4.333333 
#>   110243   110252   110274   110286   110191   110212   110245   110250 
#> 4.333333 4.000000 4.000000 4.000000 4.666667 4.000000 4.333333 6.666667 
#>   110262   110254   110269   110251   110270   110279   110200   110261 
#> 4.666667 5.333333 6.666667 5.333333 6.000000 4.666667 6.000000 6.000000 
#>   110273   110244   110258   110242   110253 
#> 6.000000 6.666667 6.333333 5.000000 6.666667