Format antibody titers.

FormatTiters formats titers into a list with one tidy data frame per viral strain

FormatTiters(titers, log2Transform = TRUE, fcMinZero = TRUE)

Arguments

titers
a data frame containing one row per subject per strain. The following columns are required:

SubjectID

Subject IDs (column name can vary)

Strain

The name of the viral strain for the observation

Pre

The pre-vaccination (or pre-infection) titer

Post

The post-vaccination (or post-infection) titer

...

Other columns which will be preserved

log2Transform
logical specifying whether titer values should be log2 transformed

fcMinZero
should negative fold changes be set to 0? Default is TRUE

Value

a list of data frames with one data frame per viral strain containing the "Pre" and "Post" titer measurements (row names are removed).

Examples

titer_list <- FormatTiters(Year1_Titers, log2Transform = TRUE, fcMinZero = TRUE)

#> - Log transforming Pre and Post columns

#> - Setting any negative log fold changes to 0