Format antibody titers.

FormatTiters formats titers into a list with one tidy data frame per viral strain

FormatTiters_OLD(titers, strains, subjectCol = "SubjectID",
  otherCols = vector(mode = "character"), d0Cols = paste0("d0_", strains),
  fcCols = paste0("fc_", strains), fcMinZero = TRUE, log2Transform = TRUE)

Arguments

titers
a data frame containing the titer information

strains
the names of the virus strains

subjectCol
the name of the column specifying a subject ID. Default is "SubjectID".

otherCols
a character vector specifying which additional columns of titers to retain. (Defaults to an empty character vector).

d0Cols
the column names of day 0 (baseline) columns

fcCols
the column names of fold change columns

fcMinZero
should negative fold changes be set to 0? Default is TRUE

log2Transform
logical specifying whether titer values should be log2 transformed

Value

a list of data frames with one data frame per viral strain containing the baseline ("d0"), fold change ("fc") and any other columns specified by the otherColumns argument.

Examples

strains <- c("A_California_7_2009", "A_Perth_16_2009", "B_Brisbane_60_2008")
titer_list <- FormatTiters(Year1_Titers, strains, subjectCol = "YaleID")

#> Error in FormatTiters(Year1_Titers, strains, subjectCol = "YaleID"): unused argument (subjectCol = "YaleID")